+++ app/src/main/java/com/hansung/sherpa/user/UserManager.kt
@@ -4,14 +4,10 @@ import android.util.Log
 import com.google.gson.Gson
 import com.hansung.sherpa.user.createuser.CreateUserRequest
 import com.hansung.sherpa.user.createuser.CreateUserResponse
-import com.hansung.sherpa.user.createuser.UserService
 import com.hansung.sherpa.user.linkpermission.LinkPermissionResponse
-import com.hansung.sherpa.user.linkpermission.LinkPermissionService
 import com.hansung.sherpa.user.login.LoginRequest
-import com.hansung.sherpa.user.login.LoginService
 import com.hansung.sherpa.user.login.User1
 import com.hansung.sherpa.user.login.UserResponse
-import com.hansung.sherpa.user.relation.GetUserRelationService
 import com.hansung.sherpa.user.relation.Relation
 import com.hansung.sherpa.user.relation.RelationResponse
 import com.hansung.sherpa.user.updateFcm.UpdateFcmRequest
@@ -56,7 +52,7 @@ class UserManager {
                         .baseUrl("http://13.209.212.166:8080/api/v1/user/")
                         .addConverterFactory(GsonConverterFactory.create())
                         .build()
-                        .create(LinkPermissionService::class.java)
+                        .create(UserService::class.java)
                         .getLinkPermissionService(caregiverEmail).execute()
                     //TODO: 잘못된 Email로 요청할 때 에러처리 해야한다.
                     result = Gson().fromJson(response.body()!!.string(), LinkPermissionResponse::class.java)
@@ -80,7 +76,7 @@ class UserManager {
                         .baseUrl("http://13.209.212.166:8080/api/v1/user/")
                         .addConverterFactory(GsonConverterFactory.create())
                         .build()
-                        .create(LoginService::class.java)
+                        .create(UserService::class.java)
                         .postLoginService(loginRequest).execute()
                     result = Gson().fromJson(response.body()!!.string(), UserResponse::class.java)
                 } catch (e:IOException){
@@ -144,8 +140,8 @@ class UserManager {
                         .baseUrl("http://13.209.212.166:8080/api/v1/userRelation/")
                         .addConverterFactory(GsonConverterFactory.create())
                         .build()
-                        .create(GetUserRelationService::class.java)
-                        .getUserRelationService(userId).execute()
+                        .create(UserService::class.java)
+                        .getRelationService(userId).execute()
                     result = Gson().fromJson(response.body()!!.string(), RelationResponse::class.java)
                 } catch(e: java.io.IOException){
                     Log.d("explain", "onFailure: 실패")
